Port dissallowed-phrases analyser to TS
